Raise the default precision from 64 to 128
`BigFloat.precision` and `BigRational.precision` are the fallback used
by every lossy operation that is called without an explicit
`precision:` argument.  64 bits is below the 53 bits of `Double` plus
any real headroom, so the out-of-the-box answer was barely better than
the built-in type these packages exist to improve on.  128 bits gives
roughly 39 decimal digits by default.

README is updated to match: the documented default, and the example
that shows how to change it -- which used to raise 64 to 128 and now
raises 128 to 256, with both outputs taken from an actual run.

Note that `RationalType.squareRoot(precision:)` and
`formSquareRoot(precision:)` keep their `Int64.bitWidth` default.  That
is a different knob -- a floor on the internal working width, on a
protocol that has no static `precision` to defer to.

+++ Sources/BigNum/BigFloat.swift
@@ -10,7 +10,7 @@ public struct BigFloat: Equatable, Hashable, Codable {  // automatic conformance
     public typealias Stride             = BigFloat
     public var scale:Exponent           // stored property
     public var mantissa:Significand     // stored property
-    public static var precision = 64
+    public static var precision = 128
     public static var roundingRule = FloatingPointRoundingRule.toNearestOrAwayFromZero
     public static var expLimit     = Self(Int16.max)
     // basic init
